created file for mentee route
