+++ logs/access.log
@@ -12,3 +12,28 @@
 ::1 - - [15/Aug/2020:08:50:33 +0000] "GET /register HTTP/1.1" 404 147
 ::1 - - [15/Aug/2020:08:55:49 +0000] "POST /mentee/register/verify HTTP/1.1" 302 38
 ::1 - - [15/Aug/2020:08:55:49 +0000] "GET /mentee/register HTTP/1.1" 200 33
+::1 - - [15/Aug/2020:09:40:00 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [15/Aug/2020:09:59:29 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [15/Aug/2020:09:59:54 +0000] "POST /mentee/register/verify HTTP/1.1" 302 38
+::1 - - [15/Aug/2020:09:59:54 +0000] "GET /mentee/register HTTP/1.1" 200 33
+::1 - - [15/Aug/2020:10:04:44 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [15/Aug/2020:10:07:47 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [15/Aug/2020:10:08:26 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [15/Aug/2020:10:09:49 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [15/Aug/2020:10:11:27 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [15/Aug/2020:10:13:43 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [15/Aug/2020:10:15:14 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [16/Aug/2020:07:04:04 +0000] "POST /mentee/register/otp HTTP/1.1" 404 159
+::1 - - [16/Aug/2020:07:04:39 +0000] "POST / HTTP/1.1" 404 140
+::1 - - [16/Aug/2020:07:04:46 +0000] "GET / HTTP/1.1" 200 41
+::1 - - [16/Aug/2020:07:04:53 +0000] "GET /register HTTP/1.1" 404 147
+::1 - - [16/Aug/2020:07:06:35 +0000] "POST /mentee/register/verify HTTP/1.1" 404 162
+::1 - - [16/Aug/2020:07:07:09 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [16/Aug/2020:07:08:21 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [16/Aug/2020:07:10:33 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [16/Aug/2020:07:10:53 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [16/Aug/2020:07:12:56 +0000] "POST /mentee/register/otp HTTP/1.1" 200 21
+::1 - - [16/Aug/2020:07:14:26 +0000] "POST /mentee/register/verify HTTP/1.1" 302 31
+::1 - - [16/Aug/2020:07:14:26 +0000] "GET /register HTTP/1.1" 404 147
+::1 - - [16/Aug/2020:07:15:34 +0000] "POST /mentee/register/verify HTTP/1.1" 302 38
+::1 - - [16/Aug/2020:07:15:34 +0000] "GET /mentee/register HTTP/1.1" 200 33
