+++ server.js
@@ -6,7 +6,7 @@ const fs = require('fs')
 var path = require('path')
 
 const routes = require('./routes/index')
-
+const menteeRoute = require('./routes/mentee')
 // Middlewares
 app.set('view engine', 'ejs')
 app.use(express.static('public'))
@@ -24,7 +24,7 @@ if (process.env._NODE_ENV === 'production') {
 
 // Routes
 app.use(routes)
-
+app.use(menteeRoute)
 
 
 // Start server
